implement generic browser
